chore(rake): Add Rakefile
Rake is a ruby implementation of Make. It is superior to Make and
integrates well with latex and python. Rake can be used to automate
tasks, like linting, compiling latex documents, cleaning temporary files
and so on. To see what you can do with rake, run `rake -T` in the root
of the project. You'll need to intall ruby first.